+++ book/src/changelog.md
@@ -1,16 +1,25 @@
 # mcpp-standard更新日志
 
+## 2025/11
+
+---
+
+**C++11 - 12 - nullptr - 指针字面量**
+
+- `Book`: [zh](https://sunrisepeak.github.io/mcpp-standard/cpp11/12-nullptr.html) / [en](https://sunrisepeak.github.io/mcpp-standard/en/cpp11/12-nullptr.html) - 2025/11/02
+- `Code`: [zh](https://github.com/Sunrisepeak/mcpp-standard/blob/main/dslings/cpp11/12-nullptr-0.cpp) / [en](https://github.com/Sunrisepeak/mcpp-standard/blob/main/dslings/en/cpp11/12-nullptr-0.cpp) - 2025/11/02
+
 ## 2025/09
 
 ---
-**C++11 - 11 - 继承造函数**
+**C++11 - 11 - 继承构造造函数**
 
 - **video:** [Bili](https://www.bilibili.com/video/BV1bspBzFEEC) / [Youtube](https://youtu.be/p7vbY8XUKnY?si=GZUn9GSW68aU94A6) / [Code](https://github.com/Sunrisepeak/mcpp-standard/blob/main/videos/cpp11/11-inherited-constructors.py)
 
 ## 2025/08
 
 ---
-**C++11 - 11 - 继承造函数**
+**C++11 - 11 - 继承构造造函数**
 
 - **[code-0](https://github.com/Sunrisepeak/mcpp-standard/blob/main/dslings/cpp11/11-inherited-constructors-0.cpp) / [code-1](https://github.com/Sunrisepeak/mcpp-standard/blob/main/dslings/cpp11/11-inherited-constructors-1.cpp) / [code-2](https://github.com/Sunrisepeak/mcpp-standard/blob/main/dslings/cpp11/11-inherited-constructors-2.cpp)** - 2025/08/29
 - **[book](https://sunrisepeak.github.io/mcpp-standard/cpp11/11-inherited-constructors.html)** - 2025/08/28